backend/requirements.txt
django==5.0.3
python-dotenv==1.0.1
django-health-check==3.16.5
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
pyyaml==6.0.1
django-cors-headers==4.3.1
django-filter==24.2
drf-extra-fields==3.7.0
drf_spectacular==0.27.2
xhtml2pdf==0.2.15
psutil==5.9.8
